requirements.txt
discord.py
requests
pymupdf
python-dotenv
langdetect
google-generativeai